altered home/search adapter for use with real API results

+++ Game-Stats-for-Dota2/app/src/main/java/com/freatnor/game_stats_for_dota2/HomeActivity.java
@@ -187,7 +187,7 @@ public class HomeActivity extends AppCompatActivity implements MatchCallback, Pl
     }
 
     @Override
-    public void playerSelected(int playerId) {
+    public void playerSelected(long playerId) {
         Intent intent = new Intent(HomeActivity.this, PlayerDetailActivity.class);
         intent.putExtra(getString(R.string.player_id_key), playerId);
         startActivity(intent);
