+++ Game-Stats-for-Dota2/app/src/main/java/com/freatnor/game_stats_for_dota2/interfaces/PlayerCallback.java
@@ -1,10 +1,8 @@
 package com.freatnor.game_stats_for_dota2.interfaces;
 
-import com.freatnor.game_stats_for_dota2.Player;
-
 /**
  * Created by Jonathan Taylor on 9/7/16.
  */
 public interface PlayerCallback {
-    void playerSelected(int playerId);
+    void playerSelected(long playerId);
 }
